Add [T13] Riwayat Kedatangan: bisa lihat riwayat parent (diri sendiri)
Added functionality to include parent customer ID in the list of customer IDs by querying the `MCustomers` table. Refactored the retrieval of customer details for `VMTAppointmentDone` objects to directly query the `MCustomers` table instead of joining with `MCustomerMembers`. Made minor formatting changes to the initialization of the `VMResponse<VMTAppointmentDone>` object.